Add github-backup.sh
Added a self-modifying script for generating a Linux script.

The first time this script is run, if an argument, the script replaces
itself with a script that runs the given argument as a Java jar.

So if passed "/path/to/github-backup-jar-with-dependencies.jar", this
github-backup.sh will become a convenient script that you can put on
your classpath to run GitHubBackup anywhere, without having to use the
fully qualified jar path or "java -jar" syntax.

All parameters to the script are piped to the jar, so the usage is
identical.